Refactor `queue` module for code size reduction
The changes consider `QUEUE_SIZE` to be a power of two, hence a
mechanism to assert such condition is required for correctness.

Use a C macro for advancing indeces, which exploits bit-wise `&`
instead of usual `%` to wrap around on index overflows.

Initialise the `tail` index to `0`, this leads to some reodering
in the flow of the `queue_enqueue` function.

Changes to be committed:
	modified:   inc/queue.h
	modified:   src/main.c
	modified:   src/queue.c
	modified:   src/queue_test.c